Module fishnet::component::prelude

commonly used imports for building components.

Re-exports

pub use super::BuildableComponent;

pub use super::ComponentState;

pub use crate::component;

pub use crate::js::ScriptType;

pub use crate::css;

Modules

routing

Routing between [Service]s and handlers.

Macros

c

add components to the page.

css

macro for generating a StyleFragment from css.

html

macro for generating Markup from html.

script

add inline js to the page

state

state_init

style

add css to the page

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

Extension

Extractor and response for extensions.

Mutex

An asynchronous Mutex-like type.

Traits

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

Render

Represents a type that can be rendered as HTML.

Type Aliases

BoxFuture

An owned dynamically typed Future for use in cases where you can’t statically type your result or need to add some indirection.

Markup

A block of markup is a string that does not need to be escaped.

Attribute Macros

component

attribute macro for creating new components

dyn_component

same as component, but forces the component to be rerendered each page visit.