Macro fishnet::css

css!() { /* proc-macro */ }

macro for generating a StyleFragment from css.

the syntax is very similar to the css syntax, with some extensions:

top level declarations

css! {
   color: red
   background-color: #00ff00;
};

will automatically be scoped to select the root class used while rendering the resulting StyleFragment

relative selectors

css! {
    > div {
        /* ... */
    }
    
    * {
        /* ... */
    }
}

all selectors are relative to the root class used while rendering the resulting StyleFragment. therefore > div will only select its direct children that are divs and * will select all children of the root class

the issue with em, ex and color hex values

due to the way rust syntax works, everytime you have a number followed by the letter e (e.g. #2effff, 1em, …), it will be interpreted as an exponential number and result in a compile error. in these special cases you can use a string literal and it will be automatically unescaped. this currently works for strings that start with a # and strings that end with em and ex, everything else will stay a string. if you want a string that matches these conditions to stay as a string, you can triple quote it:

css! {
    some-color: #2effff   /* this will not compile... */
    some-color: "#2effff" /* ...but this will! */

    some-length: 1em;     /* this will not compile... */
    some-length: "1em";   /* ...but this will! */

    some-length: 1ex;     /* this will not compile... */
    some-length: "1ex";   /* ...but this will! */

    some-other-string: "hello world!"; /* this will stay a string */
    string-with-em: """1em""";         /* this will become a normal string */
}

double class selectors and other special cases

in special cases like .root-class.other-class (e.g. both classes on the same element), you can refer to the root class using &:

css! {
    &.other-class {
        /* ... */
    }
}

this will only work if the & is the first character of the selector!