Macro fishnet::style

macro_rules! style {
    ($css: expr) => { ... };
    ($tl_class: literal, $css:expr) => { ... };
}

add css to the page

you have to call this from within a page render or it will not work.

there are two ways to use this macro:

from within a component

use fishnet::component::prelude::*;

#[component]
async fn my_component() {
    style!(css!{
        color: red;
    });

    html! {
        "hello world!"
    }
}

from outside a component

additionally to providing the style, you also have to name the top level class yourself as the first argument.

this is helpful for adding styling to functions that just return Markup.

use fishnet::{style, html, Markup, css};

async fn render_something() -> Markup {
    style!("my_class", css!{
        color: red;
    });
    
    html!{
        div class="my_class" {
            "hello world!"
        }
    }
}